update query cli
